+++ apps/web-app/package.json
@@ -35,14 +35,13 @@
     "i18next": "25.5.2",
     "i18next-browser-languagedetector": "8.2.0",
     "i18next-http-backend": "3.0.2",
-    "immer": "^10.2.0",
     "lucide-react": "0.544.0",
-    "next-themes": "^0.4.6",
+    "next-themes": "0.4.6",
     "react": "19.1.1",
     "react-dom": "19.1.1",
-    "react-i18next": "^15.7.3",
-    "recharts": "^3.2.1",
-    "sonner": "^2.0.7",
+    "react-i18next": "15.7.3",
+    "recharts": "3.2.1",
+    "sonner": "2.0.7",
     "tailwind-merge": "3.3.1",
     "tailwindcss": "4.1.13",
     "tw-animate-css": "1.3.8",
@@ -53,7 +52,7 @@
     "zustand": "5.0.8"
   },
   "devDependencies": {
-    "@types/debug": "^4.1.12",
+    "@types/debug": "4.1.12",
     "vite": "7.1.7",
     "vite-plugin-node-polyfills": "0.24.0",
     "web-vitals": "5.1.0"
