+++ apps/web-app/src/components/TransactionDialog/TransactionDialog.tsx
@@ -1,48 +1,16 @@
-import {
-  TRANSACTION_STATE,
-  txStore,
-  type SlayerTx,
-} from '@/lib/transactions/store';
-import {
-  isMessageRequest,
-  isTransactionRequest,
-  isTypedDataRequest,
-} from '@sovryn/slayer-sdk';
+import { txStore } from '@/lib/transactions/store';
 import clsx from 'clsx';
-import {
-  CircleCheckBig,
-  CircleDashed,
-  CircleX,
-  ExternalLink,
-  Loader2Icon,
-} from 'lucide-react';
-import { useCallback, useEffect, useMemo, useState, type FC } from 'react';
+import { Loader2Icon } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
-import { toast } from 'sonner';
-import { prepareTransactionRequest } from 'viem/actions';
-import {
-  useAccount,
-  useChains,
-  useConfig,
-  useSendTransaction,
-  useSignMessage,
-  useSignTypedData,
-  useSwitchChain,
-  useWaitForTransactionReceipt,
-} from 'wagmi';
-import { useStore } from 'zustand';
 import { useStoreWithEqualityFn } from 'zustand/traditional';
-import { LinkToExplorer } from '../LinkToExplorer/LinkToExplorer';
-import { Button } from '../ui/button';
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogDescription,
-  DialogFooter,
   DialogHeader,
   DialogTitle,
 } from '../ui/dialog';
+import { TxList } from './TxList';
 
 export const TransactionDialogProvider = () => {
   const { t } = useTranslation('tx');
@@ -91,332 +59,3 @@ export const TransactionDialogProvider = () => {
     </Dialog>
   );
 };
-
-const TxList = () => {
-  const { t } = useTranslation(['tx', 'common']);
-  const [isPreparing, setIsPreparing] = useState(false);
-  const { switchChain } = useSwitchChain();
-  const config = useConfig();
-  const { isConnected, chainId } = useAccount();
-
-  const items = useStoreWithEqualityFn(txStore, (state) => state.items);
-  const updateItem = useStore(txStore, (state) => state.updateItem);
-  const updateItemState = useStore(txStore, (state) => state.updateItemState);
-  const setItemError = useStore(txStore, (state) => state.setItemError);
-
-  const currentTx = useMemo(
-    () => items.find((t) => t.state !== TRANSACTION_STATE.success),
-    [items],
-  );
-
-  const currentChain = useMemo(
-    () => config.chains.find((c) => c.id === chainId),
-    [chainId, config.chains],
-  );
-  const requiredChain = useMemo(() => {
-    let id: number | undefined = undefined;
-    if (currentTx && isTransactionRequest(currentTx)) {
-      id = currentTx.request.data.chain?.id;
-    }
-
-    if (currentTx && isTypedDataRequest(currentTx)) {
-      id = currentTx.request.data.domain?.chainId
-        ? Number(currentTx.request.data.domain.chainId)
-        : undefined;
-    }
-
-    return id !== undefined
-      ? config.chains.find((c) => c.id === id)
-      : undefined;
-  }, [currentTx]);
-
-  const { signMessage, isPending: isSigning } = useSignMessage({
-    mutation: {
-      onError(error) {
-        const msg = handleErrorMessage(error);
-        toast.error(msg);
-        console.error('Message signing error', error);
-        if (!currentTx) return;
-        setItemError(currentTx.id, msg);
-      },
-      onSuccess(data) {
-        if (!currentTx) return;
-        updateItem(currentTx.id, TRANSACTION_STATE.success, {
-          transactionHash: data,
-        });
-        toast.success('Message signed successfully');
-      },
-    },
-  });
-
-  const { signTypedData, isPending: isSigningTypedData } = useSignTypedData({
-    mutation: {
-      onError(error) {
-        const msg = handleErrorMessage(error);
-        console.error('Typed Data signing error', error);
-        toast.error(msg);
-        if (!currentTx) return;
-        setItemError(currentTx?.id || '', msg);
-      },
-      onSuccess(data) {
-        if (!currentTx) return;
-        updateItem(currentTx?.id || '', TRANSACTION_STATE.success, {
-          transactionHash: data,
-        });
-        toast.success('Typed Data signed successfully');
-      },
-    },
-  });
-
-  const {
-    sendTransaction,
-    data: pendingTxHash,
-    isPending: isSending,
-  } = useSendTransaction({
-    mutation: {
-      onSettled(data, error) {
-        if (!currentTx) return;
-        if (data) {
-          updateItem(currentTx.id, TRANSACTION_STATE.pending, {
-            transactionHash: data,
-          });
-          toast.success('Transaction is broadcasted');
-        } else if (error) {
-          console.log({
-            msg: error.message,
-            cause: error.cause,
-            name: error.name,
-            stack: error.stack,
-          });
-          const msg = handleErrorMessage(error);
-          console.error('Transaction sending error', error);
-          setItemError(currentTx.id, msg);
-          toast.error(msg);
-        }
-      },
-    },
-  });
-
-  const {
-    data: receipt,
-    status: receiptStatus,
-    isLoading: isReceiptPending,
-  } = useWaitForTransactionReceipt({
-    chainId:
-      currentTx && isTransactionRequest(currentTx)
-        ? // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          (currentTx.request.data.chain?.id as any)
-        : undefined,
-    hash:
-      currentTx && isTransactionRequest(currentTx)
-        ? currentTx.res?.transactionHash
-        : undefined,
-    onReplaced: (tx) => {
-      if (!currentTx) return;
-      if (tx.reason === 'cancelled') {
-        updateItemState(currentTx.id, TRANSACTION_STATE.idle);
-        toast.error('Transaction was cancelled');
-        return;
-      }
-      updateItem(
-        currentTx.id,
-        tx.transactionReceipt.status === 'success'
-          ? TRANSACTION_STATE.success
-          : TRANSACTION_STATE.error,
-        tx.transactionReceipt,
-      );
-    },
-  });
-
-  useEffect(() => {
-    if (!currentTx || !receipt) return;
-    if (receiptStatus === 'success') {
-      updateItem(currentTx.id, TRANSACTION_STATE.success, receipt);
-    } else if (receiptStatus === 'error') {
-      updateItem(currentTx.id, TRANSACTION_STATE.error, receipt);
-      setItemError(
-        currentTx.id,
-        `Transaction failed with status: ${receipt.status}`,
-      );
-    }
-  }, [currentTx, receipt, receiptStatus]);
-
-  const handleConfirm = async () => {
-    const tx = items.find((t) => !t.res);
-    if (!tx) return;
-    try {
-      setIsPreparing(true);
-      updateItemState(tx.id, TRANSACTION_STATE.pending);
-
-      if (isMessageRequest(tx)) {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        signMessage(tx.request.data as any);
-      } else if (isTypedDataRequest(tx)) {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        signTypedData(tx.request.data as any);
-      } else if (isTransactionRequest(tx)) {
-        const prepared = await prepareTransactionRequest(
-          config.getClient(),
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          tx.request.data as any,
-        );
-        sendTransaction(prepared);
-      } else {
-        throw new Error('Unknown transaction request type');
-      }
-    } catch (e) {
-      setItemError(tx.id, handleErrorMessage(e));
-      toast.error(handleErrorMessage(e));
-      console.error('Error confirming transaction', e);
-    } finally {
-      setIsPreparing(false);
-    }
-  };
-
-  const isPending =
-    isPreparing ||
-    isSending ||
-    isSigning ||
-    isSigningTypedData ||
-    (pendingTxHash && isReceiptPending) ||
-    currentTx?.state === TRANSACTION_STATE.pending;
-
-  const confirmLabel = useMemo(() => {
-    if (currentTx) {
-      if (isMessageRequest(currentTx)) {
-        return t(($) => $.signMessage, { ns: 'tx' });
-      } else if (isTransactionRequest(currentTx)) {
-        return t(($) => $.sendTransaction, { ns: 'tx' });
-      } else if (isTypedDataRequest(currentTx)) {
-        return t(($) => $.signTypedData, { ns: 'tx' });
-      }
-    }
-    return t(($) => $.confirm, { ns: 'common' });
-  }, [currentTx]);
-
-  const handleSwitchChain = useCallback(() => {
-    if (!requiredChain) return;
-    switchChain({
-      chainId: requiredChain.id,
-      addEthereumChainParameter: {
-        nativeCurrency: requiredChain.nativeCurrency,
-        rpcUrls: requiredChain.rpcUrls.default.http,
-        chainName: requiredChain.name,
-      },
-    });
-  }, [requiredChain, switchChain]);
-
-  return (
-    <>
-      <DialogHeader>
-        <DialogTitle>{t(($) => $.title, { ns: 'tx' })}</DialogTitle>
-        <DialogDescription>
-          {t(($) => $.description, { ns: 'tx' })}
-        </DialogDescription>
-      </DialogHeader>
-      {items.map((tx, index) => (
-        <TransactionItem key={tx.id} item={tx} index={index} />
-      ))}
-
-      {!isConnected && <p>{t(($) => $.connectWallet)}</p>}
-
-      <DialogFooter>
-        <DialogClose asChild>
-          <Button variant={currentTx ? 'outline' : 'default'}>
-            {t(($) => (currentTx ? $.abort : $.continue), { ns: 'common' })}
-          </Button>
-        </DialogClose>
-
-        {currentTx && (
-          <>
-            {chainId &&
-            requiredChain !== undefined &&
-            currentChain?.id !== requiredChain?.id ? (
-              <Button onClick={handleSwitchChain}>
-                {t(($) => $.switchNetwork, {
-                  ns: 'tx',
-                  name: requiredChain.name,
-                })}
-              </Button>
-            ) : (
-              <Button onClick={handleConfirm} disabled={isPending}>
-                <Loader2Icon
-                  className={clsx(isPending ? 'animate-spin' : 'hidden')}
-                  size={16}
-                />
-                {confirmLabel}
-              </Button>
-            )}
-          </>
-        )}
-      </DialogFooter>
-    </>
-  );
-};
-
-const TransactionItem: FC<{ item: SlayerTx; index: number }> = ({
-  item,
-  index,
-}) => {
-  const isTx = isTransactionRequest(item);
-  const chains = useChains();
-  const chain = useMemo(() => {
-    if (isTx) {
-      const chainId = item.request.data.chain?.id;
-      return chains.find((c) => c.id === chainId);
-    }
-    return undefined;
-  }, [chains, isTx, item.request.data]);
-
-  return (
-    <div className="flex flex-row justify-start items-start gap-4 mb-3">
-      <div className="w-8 shrink-0 grow-0 text-center">
-        <div className="flex flex-col items-center gap-1">
-          {item.state === TRANSACTION_STATE.idle && <CircleDashed size={24} />}
-          {item.state === TRANSACTION_STATE.pending && (
-            <Loader2Icon className="animate-spin" size={24} />
-          )}
-          {item.state === TRANSACTION_STATE.success && (
-            <CircleCheckBig size={24} className="text-green-500" />
-          )}
-          {item.state === TRANSACTION_STATE.error && (
-            <CircleX size={24} className="text-red-500" />
-          )}
-          <div className="text-xs">#{index + 1}</div>
-        </div>
-      </div>
-      <div className="grow">
-        <p>{item.title}</p>
-        <p className="text-sm">{item.description}</p>
-        {isTx && chain && item.res?.transactionHash && (
-          <p className="text-sm flex flex-row justify-start gap-2 items-center mt-2">
-            <ExternalLink size={16} />
-            <LinkToExplorer chain={chain} txHash={item.res.transactionHash} />
-          </p>
-        )}
-
-        {item.error && (
-          <p className="mt-2 text-xs text-red-500">{item.error}</p>
-        )}
-      </div>
-    </div>
-  );
-};
-
-function handleErrorMessage(error: unknown): string {
-  if (error instanceof Error) {
-    if (error.cause) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const cause: any = error.cause;
-      return (
-        cause.details?.detail ??
-        cause.details ??
-        cause.shortMessage ??
-        error.message ??
-        'An unknown error occurred'
-      );
-    }
-    return error.message || 'An unknown error occurred';
-  }
-  return 'An unknown error occurred';
-}
