+++ packages/sdk/package.json
@@ -17,9 +17,9 @@
   },
   "dependencies": {
     "@sovryn/slayer-shared": "workspace:*",
-    "debug": "^4.4.3",
+    "debug": "4.4.3",
     "tslib": "2.8.1",
-    "viem": "2.37.8"
+    "viem": "2.38.6"
   },
   "devDependencies": {
     "@types/debug": "^4.1.12"
